Changed data type for balance in account.

+++ library/src/main/java/com/thoughtworks/step/Account.java
@@ -3,7 +3,7 @@ package com.thoughtworks.step;
 public class Account {
   private final String name;
   private final String number;
-  private int balance = 0;
+  private float balance = 0;
 
   public Account(String name, String number) {
     this.name = name;
@@ -18,18 +18,18 @@ public class Account {
     return number;
   }
 
-  public int getBalance() {
+  public float getBalance() {
     return balance;
   }
 
-  public int credit(int amount) throws NegAmountException {
+  public float credit(int amount) throws NegAmountException {
     if(amount <= 0) {
       throw new NegAmountException();
     }
     return balance += amount;
   }
 
-  public int debit(int amount) throws NegAmountException, InsufficientBalanceException {
+  public float debit(int amount) throws NegAmountException, InsufficientBalanceException {
     if(amount <= 0) {
       throw new NegAmountException();
     }
